@@ -1,88 +1,83 @@
 {
-  "displayName": "Meta_CAPI_Extension",
-  "name": "facebook_capi_extension",
+  "displayName": "Meta Conversions API Extension",
+  "name": "facebook_conversions_api_extension",
   "platform": "edge",
-  "version": "1.0.0",
-  "description": "Send web signal through Edge to Meta's Conversions API",
+  "version": "1.0.7",
+  "description": "Send web events to Meta's Conversions API.",
   "author": {
     "name": "Meta Platforms, Inc"
   },
   "scripts": {
-    "build": "rm -rf dist && webpack --mode=production",
-    "build:watch": "webpack --mode=development --watch=true",
-    "sandbox": "webpack --env sandbox --mode=development --watch=true",
-    "package": "npm run build && npx @adobe/reactor-packager",
-    "prepack": "npm test && npm run build",
+    "build": "webpack --mode production",
+    "build:watch": "webpack --mode development --watch",
+    "clean": "rm -rf dist",
+    "sandbox": "concurrently -k -n webpack,sandbox -c \"bgBlue,bgMagenta\" \"npm run build:watch\" \"reactor-sandbox\"",
+    "package": "npm run clean && npm run build && npx @adobe/reactor-packager",
+    "prepack": "npm run clean && npm run build",
     "lint": "eslint 'src/**'",
     "format": "pretty-quick",
     "test": "jest --coverage",
-    "test:watch": "jest --watch"
-  },
-  "husky": {
-    "hooks": {
-      "pre-commit": "pretty-quick --staged && lint-staged",
-      "pre-push": "npm test"
-    }
-  },
-  "lint-staged": {
-    "*.js": [
-      "npm run lint -- --fix"
-    ]
+    "test:watch": "jest --watch",
+    "update": "rm -rf node_modules package-lock.json && npm i",
+    "prepare": "husky install"
   },
   "dependencies": {
-    "@adobe/react-spectrum": "^3.9.0",
-    "@spectrum-icons/illustrations": "^3.2.0",
-    "@spectrum-icons/workflow": "^3.2.0",
-    "babel-polyfill": "^6.26.0",
-    "fsevents": "^2.3.2",
+    "@adobe/react-spectrum": "^3.23.0",
+    "@spectrum-icons/illustrations": "^3.5.1",
+    "@spectrum-icons/workflow": "^4.0.4",
     "js-sha256": "^0.9.0",
-    "prop-types": "^15.7.2",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
-    "react-hook-form": "^6.15.5"
+    "prop-types": "^15.8.1",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "react-hook-form": "^7.39.5",
+    "react-stately": "^3.19.0"
   },
   "devDependencies": {
     "@adobe/reactor-sandbox": "^12.4.0",
-    "@babel/core": "^7.13.15",
-    "@babel/plugin-proposal-class-properties": "^7.12.1",
-    "@babel/plugin-transform-runtime": "^7.13.15",
-    "@babel/preset-env": "^7.13.15",
-    "@babel/preset-react": "^7.13.13",
-    "@babel/runtime": "^7.12.5",
-    "@testing-library/jest-dom": "^5.11.10",
-    "@testing-library/react": "^11.2.6",
-    "babel-eslint": "^10.1.0",
-    "babel-jest": "^26.6.3",
-    "babel-loader": "^8.1.0",
-    "camelcase": "^6.2.0",
-    "capitalize": "^2.0.3",
-    "css-loader": "^5.2.0",
-    "eslint": "^7.23.0",
-    "eslint-config-airbnb": "^18.2.1",
-    "eslint-config-prettier": "^6.15.0",
-    "eslint-plugin-import": "^2.22.1",
-    "eslint-plugin-jest-dom": "^3.7.0",
-    "eslint-plugin-jsx-a11y": "^6.4.1",
-    "eslint-plugin-prettier": "^3.1.4",
-    "eslint-plugin-react": "^7.23.2",
-    "eslint-plugin-react-hooks": "^4.2.0",
-    "eslint-plugin-testing-library": "^3.10.2",
+    "@babel/core": "^7.20.2",
+    "@babel/eslint-parser": "^7.19.1",
+    "@babel/plugin-proposal-class-properties": "^7.18.6",
+    "@babel/plugin-syntax-import-assertions": "^7.20.0",
+    "@babel/plugin-transform-runtime": "^7.19.6",
+    "@babel/preset-env": "^7.20.2",
+    "@babel/preset-react": "^7.18.6",
+    "@babel/runtime": "^7.20.1",
+    "@testing-library/jest-dom": "^5.16.5",
+    "@testing-library/react": "^13.4.0",
+    "@testing-library/user-event": "^14.4.3",
+    "babel-jest": "^29.3.1",
+    "babel-loader": "^9.1.0",
+    "camelcase": "^7.0.0",
+    "capitalize": "^2.0.4",
+    "concurrently": "^7.6.0",
+    "css-loader": "^6.7.2",
+    "eslint": "^8.28.0",
+    "eslint-config-airbnb": "^19.0.4",
+    "eslint-config-prettier": "^8.5.0",
+    "eslint-plugin-header": "^3.1.1",
+    "eslint-plugin-import": "^2.26.0",
+    "eslint-plugin-jest-dom": "^4.0.3",
+    "eslint-plugin-jsx-a11y": "^6.6.1",
+    "eslint-plugin-prettier": "^4.2.1",
+    "eslint-plugin-react": "^7.31.11",
+    "eslint-plugin-react-hooks": "^4.6.0",
+    "eslint-plugin-testing-library": "^5.9.1",
     "file-loader": "^6.2.0",
-    "fs-extra": "^9.0.1",
-    "html-webpack-plugin": "^4.5.0",
-    "husky": "^4.3.0",
-    "jest": "^26.6.3",
-    "jest-runner-eslint": "^0.10.0",
-    "lint-staged": "^10.5.1",
-    "prettier": "^2.1.2",
-    "pretty-quick": "^3.1.0",
-    "style-loader": "^2.0.0",
-    "stylus": "^0.54.8",
-    "stylus-loader": "^4.2.0",
+    "fs-extra": "^10.1.0",
+    "html-webpack-plugin": "^5.5.0",
+    "husky": "^8.0.2",
+    "jest": "^29.3.1",
+    "jest-environment-jsdom": "^29.3.1",
+    "jest-runner-eslint": "^1.1.0",
+    "lint-staged": "^13.0.3",
+    "prettier": "^2.7.1",
+    "pretty-quick": "^3.1.3",
+    "style-loader": "^3.3.1",
+    "stylus": "^0.59.0",
+    "stylus-loader": "^7.1.0",
     "url-loader": "^4.1.1",
-    "webpack": "^5.64.4",
-    "webpack-cli": "^4.9.1",
-    "webpack-shell-plugin": "^0.5.0",
-    "yargs": "^16.1.0"
+    "webpack": "^5.75.0",
+    "webpack-cli": "^5.0.0",
+    "yargs": "^17.6.2"
   }
 }
