@@ -1,10 +1,19 @@
-/**
- * (c) Meta Platforms, Inc. and affiliates. Confidential and proprietary.
- *
- * @format
- */
+/*
+Copyright 2022 Adobe. All rights reserved.
+This file is licensed to you under the Apache License, Version 2.0 (the "License");
+you may not use this file except in compliance with the License. You may obtain a copy
+of the License at http://www.apache.org/licenses/LICENSE-2.0
+
+Unless required by applicable law or agreed to in writing, software distributed under
+the License is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR REPRESENTATIONS
+OF ANY KIND, either express or implied. See the License for the specific language
+governing permissions and limitations under the License.
+*/
+
+/* eslint-disable jsx-a11y/anchor-is-valid */
+
 import React from 'react';
-import {useFormContext} from 'react-hook-form';
+import { useFormContext } from 'react-hook-form';
 import {
   Checkbox,
   Content,
@@ -15,34 +24,30 @@ import {
   Text,
   TextField,
   TextArea,
+  ContextualHelp
 } from '@adobe/react-spectrum';
 import WrappedTextField from '../../components/wrappedTextField';
 import WrappedCheckboxComponent from '../../components/wrappedCheckboxComponent';
+import WrappedComboboxField from '../../components/wrappedComboBox';
 import Parameters from './getParameters';
+import actionSources from '../../utils/actionSources';
 
-export default () => {
-  const {watch} = useFormContext();
-  const isTestEvent = watch('isTestEvent');
-  const {customerInformation, serverEvents} = Parameters();
-  const parametersURI =
-    'https://developers.facebook.com/docs/marketing-api/conversions-api/parameters';
+export default function SendEventSectionFields() {
+  const { watch } = useFormContext();
+  const [isTestEvent, actionSource] = watch(['isTestEvent', 'actionSource']);
+
+  const { customerInformation, serverEvents } = Parameters({ actionSource });
+
   const dpoURI =
     'https://developers.facebook.com/docs/marketing-apis/data-processing-options/';
-  const serverEventsURI =
-    'https://developers.facebook.com/docs/marketing-api/conversions-api/parameters/server-event';
   const customerInformationURI =
     'https://developers.facebook.com/docs/marketing-api/conversions-api/parameters/customer-information-parameters';
   const customDataURI =
     'https://developers.facebook.com/docs/marketing-api/conversions-api/parameters/custom-data';
 
-  const facebookEventSetupText = `Use the data mapping below to configure a Meta event using
-    your data from Adobe Edge. These events will be sent to your Pixel via Meta Conversions
-    API. For more information about these parameters, go to `;
-  const serverEventsParametersText =
-    'Send actions that occur as a Meta Standard or Custom Event. For more detail, see the ';
   const customerInformationText = `These parameters are a set of identifiers Meta can use
     for targeted attribution. You must provide at least one of the following keys in your
-    request. For more detail, see the `;
+    request. For more details, see the `;
   const customDataText = `Use these parameters to send additional data we can use for ads
     delivery optimization. For Purchase events, value and currency are required. If data elements
     are being utilized, please validate that they are being replaced correcly. For more
@@ -54,57 +59,94 @@ export default () => {
   return (
     <Flex direction="row" gap="size-200">
       <Flex direction="column" gap="size-65">
-        <Heading> Meta Event Setup </Heading>
-        <Divider size="S" />
         <Content>
-          <Text>{facebookEventSetupText}</Text>
+          <Text>
+            Use the data mapping below to configure a Meta event using your data
+            from Adobe Edge. These events will be sent to your Pixel via Meta
+            Conversions API. For more information about these parameters, go to
+          </Text>{' '}
           <Link>
-            <a href={parametersURI} target="_blank" rel="noreferrer">
+            <a
+              href="https://developers.facebook.com/docs/marketing-api/conversions-api/parameters"
+              target="_blank"
+              rel="noreferrer"
+            >
               Meta for Developers.
             </a>
           </Link>
         </Content>
 
-        <Heading marginTop="1em"> Server Event Parameters</Heading>
-        <Divider size="S" />
-        <Content>
-          <Content marginBottom="1em">
-            <Text>{serverEventsParametersText}</Text>
-            <Link>
-              <a href={serverEventsURI} target="_blank" rel="noreferrer">
-                documentation.
-              </a>
-            </Link>
-          </Content>
-          {serverEvents.map(([name, label], index) => {
+        <Flex alignItems="center" gap="size-75">
+          <Heading level="3"> Server Event Parameters</Heading>
+
+          <ContextualHelp>
+            <Heading>Need help?</Heading>
+            <Content>
+              Send actions that occur as a Meta Standard or Custom Event. For
+              more details, see the{' '}
+              <Link>
+                <a
+                  href="https://developers.facebook.com/docs/marketing-api/conversions-api/parameters/server-event"
+                  rel="noreferrer"
+                  target="_blank"
+                >
+                  documentation
+                </a>
+              </Link>
+              .
+            </Content>
+          </ContextualHelp>
+        </Flex>
+
+        <WrappedComboboxField
+          minWidth="size-4600"
+          width="size-4600"
+          name="actionSource"
+          label="Action Source"
+          necessityIndicator="label"
+          description="This field allows you to specify where your conversion occurred."
+          isRequired
+          allowsCustomValue
+          defaultItems={actionSources
+            .getActionSourceNames()
+            .map((q) => ({ id: q, name: q }))}
+        />
+
+        {serverEvents.map(
+          ([name, label, description, isRequired, contextualHelp]) => {
             return (
               <WrappedTextField
-                key={index}
+                key={name}
                 name={name}
                 component={TextField}
                 width="size-4600"
                 label={label}
+                description={description}
+                isRequired={isRequired}
+                necessityIndicator="label"
+                contextualHelp={contextualHelp}
                 supportDataElement
               />
             );
-          })}
-          <WrappedCheckboxComponent component={Checkbox} name="lduEnabled">
-            <Text>Enable Limited Data Use</Text>
-          </WrappedCheckboxComponent>
-          <Content>
-            <Text>
-              Learn more about{' '}
-              <Link>
-                <a href={dpoURI} target="_blank" rel="noreferrer">
-                  Limited Data Use and Data Processing Options
-                </a>
-              </Link>
-            </Text>
-          </Content>
+          }
+        )}
+        <WrappedCheckboxComponent component={Checkbox} name="lduEnabled">
+          <Text>Enable Limited Data Use</Text>
+        </WrappedCheckboxComponent>
+
+        <Content>
+          <Text>
+            Learn more about{' '}
+            <Link>
+              <a href={dpoURI} target="_blank" rel="noreferrer">
+                Limited Data Use and Data Processing Options
+              </a>
+            </Link>
+          </Text>
         </Content>
 
-        <Heading marginTop="1em"> Customer Information Parameters </Heading>
-        <Divider size="S" />
+        <Heading level="3"> Customer Information Parameters </Heading>
+
         <Content>
           <Content marginBottom="1em">
             <Text>{customerInformationText}</Text>
@@ -114,10 +156,10 @@ export default () => {
               </a>
             </Link>
           </Content>
-          {customerInformation.map(([name, label], index) => {
+          {customerInformation.map(([name, label]) => {
             return (
               <WrappedTextField
-                key={index}
+                key={name}
                 name={name}
                 component={TextField}
                 width="size-4600"
@@ -171,4 +213,4 @@ export default () => {
       </Flex>
     </Flex>
   );
-};
+}
