@@ -1,13 +1,21 @@
-/**
- * (c) Meta Platforms, Inc. and affiliates. Confidential and proprietary.
- */
+/*
+Copyright 2022 Adobe. All rights reserved.
+This file is licensed to you under the Apache License, Version 2.0 (the "License");
+you may not use this file except in compliance with the License. You may obtain a copy
+of the License at http://www.apache.org/licenses/LICENSE-2.0
+
+Unless required by applicable law or agreed to in writing, software distributed under
+the License is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR REPRESENTATIONS
+OF ANY KIND, either express or implied. See the License for the specific language
+governing permissions and limitations under the License.
+*/
 
 /* istanbul ignore file */
 
 import React from 'react';
 import { useFormContext } from 'react-hook-form';
 
-export default () => {
+export default function DisplayFormState() {
   const { formState } = useFormContext();
   return (
     <div style={{ margin: '1rem 0' }}>
@@ -23,4 +31,4 @@ export default () => {
       </pre>
     </div>
   );
-};
+}
