@@ -1,6 +1,14 @@
-/**
- * (c) Meta Platforms, Inc. and affiliates. Confidential and proprietary.
- */
+/*
+Copyright 2022 Adobe. All rights reserved.
+This file is licensed to you under the Apache License, Version 2.0 (the "License");
+you may not use this file except in compliance with the License. You may obtain a copy
+of the License at http://www.apache.org/licenses/LICENSE-2.0
+
+Unless required by applicable law or agreed to in writing, software distributed under
+the License is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR REPRESENTATIONS
+OF ANY KIND, either express or implied. See the License for the specific language
+governing permissions and limitations under the License.
+*/
 
 import React, { Component } from 'react';
 import ErrorMessage from './errorMessage';
@@ -16,7 +24,7 @@ export default class ErrorBoundary extends Component {
     return { error };
   }
 
-  componentDidCatch(/*error, info*/) {
+  componentDidCatch(/* error, info */) {
     // You can also log the error to an error reporting service
     // console.error(error, info);
   }
