@@ -1,15 +1,25 @@
-/**
- * (c) Meta Platforms, Inc. and affiliates. Confidential and proprietary.
- */
+/*
+Copyright 2022 Adobe. All rights reserved.
+This file is licensed to you under the Apache License, Version 2.0 (the "License");
+you may not use this file except in compliance with the License. You may obtain a copy
+of the License at http://www.apache.org/licenses/LICENSE-2.0
+
+Unless required by applicable law or agreed to in writing, software distributed under
+the License is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR REPRESENTATIONS
+OF ANY KIND, either express or implied. See the License for the specific language
+governing permissions and limitations under the License.
+*/
 
 import React from 'react';
 import { IllustratedMessage, Heading, Content } from '@adobe/react-spectrum';
 import NotFound from '@spectrum-icons/illustrations/NotFound';
 
-export default ({ message = 'There was an error!' }) => (
-  <IllustratedMessage marginTop="size-1000">
-    <NotFound />
-    <Heading>An error has occured</Heading>
-    <Content>{message}</Content>
-  </IllustratedMessage>
-);
+export default function ErrorMessage({ message = 'There was an error!' }) {
+  return (
+    <IllustratedMessage marginTop="size-1000">
+      <NotFound />
+      <Heading>An error has occured</Heading>
+      <Content>{message}</Content>
+    </IllustratedMessage>
+  );
+}
