@@ -1,16 +1,21 @@
-/**
- * (c) Meta Platforms, Inc. and affiliates. Confidential and proprietary.
- */
+/*
+Copyright 2022 Adobe. All rights reserved.
+This file is licensed to you under the Apache License, Version 2.0 (the "License");
+you may not use this file except in compliance with the License. You may obtain a copy
+of the License at http://www.apache.org/licenses/LICENSE-2.0
+
+Unless required by applicable law or agreed to in writing, software distributed under
+the License is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR REPRESENTATIONS
+OF ANY KIND, either express or implied. See the License for the specific language
+governing permissions and limitations under the License.
+*/
 
 /* eslint-disable react/jsx-props-no-spreading */
 
 import React from 'react';
-import { useFormContext, Controller } from 'react-hook-form';
-import { ActionButton } from '@adobe/react-spectrum';
-import Data from '@spectrum-icons/workflow/Data';
-import ValidationWrapper from './validationWrapper';
+import { Controller } from 'react-hook-form';
 
-export default ({
+export default function WrappedCheckboxComponent({
   name: componentName,
   component: Component,
   onChange: componentOnChange,
@@ -18,7 +23,7 @@ export default ({
   supportDataElement,
   defaultValue = '',
   ...rest
-}) => {
+}) {
   return (
     <Controller
       name={componentName}
@@ -26,10 +31,10 @@ export default ({
       render={({ onChange, value, name, ref }) => (
         <Component
           name={name}
-          onChange={(e) =>{
+          onChange={(e) => {
             onChange(e);
-            if (componentOnChange){
-              componentOnChange(e)
+            if (componentOnChange) {
+              componentOnChange(e);
             }
           }}
           isSelected={value}
@@ -40,4 +45,4 @@ export default ({
       {...rest}
     />
   );
-};
+}
