@@ -1,21 +1,24 @@
-/**
- * (c) Meta Platforms, Inc. and affiliates. Confidential and proprietary.
- */
+/*
+Copyright 2022 Adobe. All rights reserved.
+This file is licensed to you under the Apache License, Version 2.0 (the "License");
+you may not use this file except in compliance with the License. You may obtain a copy
+of the License at http://www.apache.org/licenses/LICENSE-2.0
+
+Unless required by applicable law or agreed to in writing, software distributed under
+the License is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR REPRESENTATIONS
+OF ANY KIND, either express or implied. See the License for the specific language
+governing permissions and limitations under the License.
+*/
 export default (values) => {
   const errors = {};
 
   if (!values.pixelId) {
-    errors.pixelId = {
-      message: 'Please specify your Meta Pixel ID',
-      type: 'required'
-    };
+    errors.pixelId = 'Please specify your Meta Pixel ID.';
   }
 
   if (!values.accessToken) {
-    errors.accessToken = {
-      message: 'Please specify your access token',
-      type: 'required'
-    };
+    errors.accessToken = 'Please specify your access token.';
   }
+
   return errors;
 };
